backend/requirements.txt
fastapi
minio
passlib[bcrypt]
pydantic
PyJWT
pymysql
python-dotenv
python-jose[cryptography]
python-multipart
SQLAlchemy
uvicorn[standard]